syntax: make IsIncomplete more robust
It now works with Parser.Words reliably. It used to rely on counting
the number of open statements, and there are none there.
Moreover, the way it used the current quote state to check
for missing closing tokens was flawed, as it did not work for
closing double quotes or curly braces for parameter expansions.

Fixes #937.