interp: initial support for $! and wait PID
Background shells are implemented as goroutines, not as forked
processes with real PIDs, so we instead use virtual PIDs
with a "g" prefix, like g1, g2, g3, etc.

The main benefit here is to be able to wait for background commands
and collect their status codes, being able to react if any of them fail.
"wait" with no arguments always succeeds.

Fixes #221.