interp: make a full copy of the environment for background subshells
If a parent shell sets a variable once a background subshell is running,
the subshell should not see that change, nor should the change cause
any data races. We allowed the parent and child shells to share
a map of variables when the child was a background subshell,
which caused one of the added tests to fail often:

    --- FAIL: TestRunnerRun/#456 (0.00s)
        interp_test.go:3662: wrong output in "{ for n in {0..9}; do { echo $n; } & done; wait; } | sort":
            want: "0\n1\n2\n3\n4\n5\n6\n7\n8\n9\n"
            got:  "9\n9\n9\n9\n9\n9\n9\n9\n9\n9\n"

Moreover, the newly added background shell tests would cause a number of
"concurrent map read and write" and data race detector panics too.

overlayEnviron can be clever in sharing a variables map with the parent
only when the subshell is not being run in the background.
When the subshell is intended for background (concurrent) use,
we need to make a full copy of the variables to avoid races.