_js: drop mvdan-sh JS package built with GopherJS
I don't have the time to properly maintain a JS port of this library,
or to learn enough JS to be a good steward for it.
Moreover, I built mvdan-sh on top of GopherJS, which is still somewhat
maintained, but is struggling to support the latest versions of Go -
which this library uses.

The good news is that someone else built a new JS wrapper library,
which is actively maintained and uses WASM instead of GopherJS,
bringing a speed-up of over 4x: https://github.com/un-ts/sh-syntax

For now, given that I don't intend to further develop mvdan-sh,
I am closing existing issues and deleting the code,
but leaving notes in the README and the npm package's page
so that any users get pointed in the right direction.

Fixes #1145.