interp: expose ExitStatus and deprecate NewExitStatus and IsExitStatus
I wrote this error logic with xerrors.As back in 2019, and rather than
exposing the type and letting the user interact with errors.As directly,
I only exposed narrower APIs.

I'm honestly not sure why I did that; these two APIs are more surface
and complexity, and they also prevent some flexibility.
For example, I might want to create a custom error which satisfies
errors.As with interp.ExitStatus by implementing my own As method.
That's only possible if the ExitStatus type is actually exposed to me.