interp: subshells in a pipe are run concurrently
I randomly ran into a `go test -race` failure in the interp package
caused by this new test case added recently:

    GOSH_CMD=exit_5 $GOSH_PROG | GOSH_CMD=exit_0 $GOSH_PROG

The issue was that we were not making a full copy of the environment
when creating a subshell for concurrent execution within a pipe.
So, if both ends of a shell pipe modified the environment such
as in this case here, we would potentially cause a data race.

Add a test which is way more likely to spot these races in the future,
and fix the data race by making a full copy of the environment.