expand: always use pattern.NoGlobStar option
The expand package has direct support for `**` globs, so does not need
to rely on the support in the pattern package. By disabling NoGlobStar
we make it less likely that bugs or slight differences in support for
`**` in one package will affect the other.

This commit also adds some tests demonstrating that the GlobStar support
in the expand package works.