pattern: add test cases for matching dot-prefixed filenames
In both POSIX and Bash, * and ** do not match dot-prefixed filenames
by default, such that `*/foo` matches `dir/foo` but not `.dir/foo`.

We implemented these semantics correctly in the expand package
via its "matchHidden" logic, as it advanced each path element
when performing glob expansion.

However, direct users of the pattern package with the Filenames mode
do not see the correct semantics in terms of leading dots.
The tests are added here, and the fix is coming in a later commit.