pattern: remove the broken Braces mode
It was added years ago with the logic of giving some brace expansion
functionality to users of the pattern package without needing to use
the expand package.

However, this seems flawed for a number of reasons:

1) Brace expansion and pattern matching are entirely different features.
   The former expands a field into many, and the latter tends to perform
   some action based on string matches, such as expanding filenames
   when globbing, or returning a boolean like `[[ foo == pattern ]]`.

2) Brace expansion is meant to expand one string field into many,
   hence its API needs to return a list of new fields or words,
   much like the existing expand.Braces function.
   An API returning a regular expression is not helpful.

3) The interp and expand packages did not make use of this feature,
   as the expand package already performs brace expansion properly
   with the help of the syntax.SplitBraces API.

Given that it was a flawed feature, that we didn't use it at all,
and that I can't find a single user when searching github.com,
remove the feature entirely for now.

This is technically a breaking change, but I think it's reasonable
to break APIs when they are clearly broken for any existing user.
If any existing users exist, and may read this commit message,
they can file an issue if this is a problem and we can discuss
next steps once we learn how they were using the API.

HasMeta and QuoteMeta no longer need a Mode parameter now,
which makes sense, given that the set of special characters used
for pattern matching are fixed.