fileutil: treat all non-regular files as definitely not shell scripts
As Kir Kolyshkin correctly points out, the presence of a named pipe file
which looks like it could contain shell could lead to a hang.
This is because shfmt reads from such a file to look for a shebang
or to format the file directly if it has a shell filename extension.

Just like we already do for symlinks, which are non-regular files,
refuse to do anything with any other type of non-regular file
when recursively formatting directories.

This is an alternative version of Kir's PR at #1089, as we want
the fix to be in the fileutil package. The added test also hangs
without the fix, so it acts as a good regression test.

Closes #1089.