interp: clear exitStatus.err after evaluating conditions
Otherwise, "if" or "while" conditions which execute external programs
would set "err" to a Go error, and we would only clear "code",
leading to the error still being surfaced to the user:

    --- FAIL: TestRunnerRun/#259 (0.00s)
        interp_test.go:3752: input: "while GOSH_CMD=print_fail $GOSH_PROG; do echo foo_interp_missing; done"
        interp_test.go:3783: wrong output in "while GOSH_CMD=print_fail $GOSH_PROG; do echo foo_interp_missing; done":
            want: "exec fail\n"
            got:  "exec fail\nexit status 1"
    --- FAIL: TestRunnerRun/#250 (0.00s)
        interp_test.go:3752: input: "if GOSH_CMD=print_fail $GOSH_PROG; then echo foo_interp_missing; fi"
        interp_test.go:3783: wrong output in "if GOSH_CMD=print_fail $GOSH_PROG; then echo foo_interp_missing; fi":
            want: "exec fail\n"
            got:  "exec fail\nexit status 1"# Please enter the commit message for your changes. Lines starting

This is a recent regression, which we did not spot as we had no tests
for "if" or "while" with conditions calling external programs.
The changes which introduced the problem added the ability for failing
commands to carry a Go error value, as that can include extra info
which the Go API user might find useful.
However, this bit of code was not adapted accordingly.

Fixes #1182.