syntax: fix "esac" as first pattern match literal
when a case statement has literal "esac" as its first pattern, it should be preceded by a left
parenthesis. the parser already permits this, but not the syntax printer.

this change adds the conclusion that the only way for "esac" to pass the parser as a first
pattern matching literal is if it was indeed preceded by a left parenthesis.