interp: print output error if `cd` fails
`cd` currently just exits with status 1 without printing any output.
This can be confusing when using inside `gosh` for example, because
it's not clear that an error happened as there is no feedback for
non-zero statuses, and not other error was print.

This patch replicates the behavior of Bash, outputing
"no such file or directory" or "permission denied".