syntax: add iterator versions of Stmts and Interactive with a Seq suffix
Using the standard iterator type means that they can be used
in a range-over-func loop, which is a pretty neat user experience.

Note that the iterator versions return an error at each yield,
allowing the caller to decide whether or not to stop on errors
or to collect multiple errors rather than just the first one.