expand: Config.ProcSubst is not a work in progress
I think I wrote this comment when I did the first half of #451,
but then I did the second half of that issue the very next day
and it looks like I forgot to remove this note.

In any case, all of that happened in late 2019.
It has been more than five years since, and we are still at a v3,
so it is extremely unlikely that we will change the API now.
Not before v4, anyway.