interp: shorten the shopt error on unsupported options
Showing the on/off values for the default does not make much sense
when shown to the user, especially because they might be trying
to turn the option on or off depending on the use of the -s or -u flags.

We don't support setting unsupported options to anything,
so just don't mention any values at all.