interp: fix POSIX and Bash option separation in shopt
When using shopt without the -o flag, we would attempt to support
both POSIX and Bash options, when in fact only Bash options should work,
as the -o flag switches the command to work on POSIX options instead.

In fact, `shopt pipefail`, which attempts to read a POSIX option
without using the `-o` flag, would cause a panic due to the buggy code.

Add tests, and fix the code to flip between the two modes correctly.