syntax: tokenize parameter expansions rune by rune
This replaces paramExpName, and makes the code a bit easier to follow
as we have to fit less logic and edge cases into quote states.
Parameter expansions are complex, and they'll just get more complicated
with the addition of zsh support, so it's best to keep the parser
in full control of exactly how they are parsed.