syntax: unify parameter expansion parsing
We used to handle short parameter expansions like $foo separately
from the full parameter expansions like ${#foo} or ${foo[1]}.
Handle them all in the same paramExp method, which will be useful
as we add support for more language features in the future.

While here, document the method a bit better.