syntax: improve language checks with langErr via checkLang
Rather than wrapping almost every use of langErr with a manual
language check conditional, do that in checkLang.

On top of that, stop assuming that LangPosix is the opposite to
"bash-like or mksh", because that will stop being true soon enough
when we add more languages.

Similarly, Lang.isBash() is now Lang.is(LangBash).