syntax: turn LangVariant into a bitset
This should be an almost entirely backwards compatible change,
given that the zero value of LangVariant is still accepted by the API
to mean LangBash.

The only breakage is if any user code assumed that, for example:

    var zero LangVariant
    assert(zero == LangBash)

That will no longer be true, because LangBash is now 1.
However, the docs never guaranteed anything about the zero value
or any of the public API values, so I think this should be OK.

Most importantly, being able to use LangVariant as a bitset
is extremely useful, particularly for the parser checking for features.