all: add initial Zsh support
Add all the machinery to both the syntax parser and the shfmt tool.

The only feature we add for Zsh in this first commit is ${+foo},
mainly to sanity check that the language is plumbed through correctly.

More features of the Zsh language will follow after this commit.
Note that we're not starting from Bash as a base for Zsh,
even though they share some features, to be conservative in only adding
the features that we can actually know should be supported.

Given how complex Zsh is - even more so than Bash - it seems best
to start from a POSIX base and add features as we go.

Updates #120.