syntax: track whether the Parser read io.EOF via a boolean field
Tracking this via either the `bs` or `r` fields was very confusing.
`bs == nil` signaling EOF is ambiguous, because the parser starts
with bs being nil before it has read anything.
And `r == utf8.RuneSelf` means that the next rune is past EOF,
not that we have already read EOF, which could be many bytes later.