syntax: print the new ParamExp.Flags field
This was missed in the introduction of the field
because the logic in TestPrintFiles would incorrectly skip file tests
which are not valid when parsed as Bash.

We also missed the field in Walk. In turn, this caused the parser tests
to not notice that the end of the literal was off by one.

Fixes #1213.