cmd/shfmt: exit with a non-zero status when -l prints any filenames
This matches the change that was already made to the -d flag,
and aligns shfmt with other formatters like prettier and rustfmt.

Any existing users of `shfmt -l` who don't want this change
can always choose to ignore the exit status code when any filenames
are printed to stdout.

Fixes #1223.