expand,interp: do not force env var names on Windows to be uppercase
This is OK when all we do is get and set specific env vars,
but when listing env vars or passing them to external programs,
we should aim to keep the original case of env var names.

Note that the work here is not done, as overlayEnviron does not support
case-insensitive env vars, hence the remaining TODOs.

Fixes #1232.