interp: record original env var names in overlayEnviron
The previous commit fixed the expand package so that it doesn't assume
that all env var names should be uppercased on Windows.

However, as shown by the TODOs in the recently added tests,
any variables set via the interpreter itself did not properly apply
the case insensitive rules on Windows. Do that.

See #1232.